Rename authClient in examples
