+++ examples/demo/src/authProvider.js
