+++ examples/graphcool-demo/src/authProvider.js
