+++ examples/simple/authProvider.js
